requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
sqlalchemy[asyncio]
asyncpg
pydantic
google-genai